stats_streamlit_db/requirements.txt
streamlit==1.37.1
psycopg2-binary==2.9.9
pandas==2.2.2
passlib
bcrypt==3.2.2